backend/requirements.txt
flask==3.1.*
flask-cors==5.*
psycopg2-binary==2.9.*
bcrypt==4.*
PyJWT==2.*
python-dotenv==1.*
anthropic==0.78.*
openai==2.17.*
google-genai==1.33.*
python-docx==1.*
